// src/types/components/type.ts
export interface IRoomInfo {
	id?: string;
	src?: string;
	title?: string;
	address?: string;
	price?: string;
	roomType?: string;
	bedRoom?: string;
	sale?: string;
	acreage?: string;
	des?: string;
	limitNumberOfPeople?: string;
	bathRoom?: string;
	createdAt?: string;
}

export interface INews {
	id?: string;
	img?: string;
	des?: string;
	title?: string;
	createdAt?: string;
	updatedAt?: string;
}